On the ra-git branch: To avoid having to reimplement quite a bit of the repos layer code to support a git backend (as the branch currently does), start moving a bit of git code to the fs layer so the existing repos layer code can be used to access git state.
For this experiment I try to build a readonly git filesystem to just support
libsvn_ra_git's access to read the repository.

As a side effect this filesystem may support other code like 'svnadmin dump',
to work on this level as this happens to have the same requirements. It might
also make it easy to publish git repositories on a Subversion compatible url.

The backend for this filesystem will be a git repository
(accessed by libgit2) and an sqlite database containing some mapping
information.

The short term goal is to support svn:externals pointing to a git repository,
like the current branch does.

* build.conf
  (libsvn_fs_git): New filesystem.
  (libsvn_ra_git,
   fs-libs): Declare usage of libsvn_fs_git.

* build/generator/gen_win_dependencies.py
  (_find_libgit2): Use revision instead of PATCH. Define
    SVN_LIBSVN_FS_LINKS_FS_GIT.

* subversion/include/svn_fs.h
  (SVN_FS_TYPE_GIT): Define.

* subversion/libsvn_fs/fs-loader.c
  (git_defn): New definition.
  (base_defn): Hook definition.
  (synchronized_initialize): Reset the new last filesystem.

* subversion/libsvn_fs/fs-loader.h
  (fs includes): Add ../libsvn_fs_git/fs_init.h.

* subversion/libsvn_fs_git
  New directory.

* subversion/libsvn_fs_git/fs.c
  New file.
* subversion/libsvn_fs_git/fs_git.h
  New file.
* subversion/libsvn_fs_git/fs_init.h
  New file.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/ra-git@1717253 13f79535-47bb-0310-9956-ffa450edef68