On the parallel-put branch: Actually allow concurrent puts in FSFS.
The idea is to simply stream into a svn_spillbuf_t and only to append
to the protorev file when the stream gets closed, i.e. when it is
completed.  We only need to serialize the last part and can allow any
number of text reps to be built up in parallel.

There is also no need to check for the same text being written twice
concurrently because that is semantically undefined anyway.  Our code
will let the last one win and prevents corruption. 

Note that this is not safe, yet, as prop sets and making the tree
node mutable have not been mutexed, yet.

* subversion/libsvn_fs_fs/transaction.c
  (rep_write_baton): Add spill buffer as an alternative to the file.
  (rep_write_open_buffer): New function similar to rep_write_open_file.
  (rep_write_get_baton): In concurrent mode, use the spill buffer.
  (rep_write_contents_close): Same. Also, in concurrent mode, we need
                              finally open the protorev file and copy
                              our data into it.  Don't lose updates
                              caused by concurrent prop sets.

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/parallel-put@1719905 13f79535-47bb-0310-9956-ffa450edef68