As suggested during the Berlin 2016 hackathon, this is a how-hard-can-it-be hack that removes the need to store pristines permanently.
None of its parts are in a usable state.  It is
simply there to demonstrate what kind of code
changes would be required, such that others may
find inspiration and implement a proper design
elsewhere.  Hence, no specifics on the changes
are given here, just a simple dump.

Using ra_svn, 17 Tests are failing, mainly around
conflict scenarios that use the pristine as an
implicit store.

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/pristineless-hack@1769826 13f79535-47bb-0310-9956-ffa450edef68