On the addremove branch, introduce new a API for creating local moves.
This new API works on deleted and added nodes which are present in the
working copy. It makes it possible to change copyfrom info of added or
copied nodes to point at a deleted node, so they become part of a local move.

The heuristic which matches up deletes and adds is not documented.
For now, it only works for files and relies on the pristine SHA1 checksum
to determine identify. In the future, we may add additional heuristics,
including repository log scanning as implemented in the conflict resolver.

While this feature was invented for addremove, it could eventually be useful
outside the context of addremove. For instance, merges currently produce
copyfrom which points to a different branch if a move is merged without
a tree conflict (see issue #2685). Perhaps, in the future, a merge could
use this new API to create moves in the working copy as a post-processing
step, fixing up copyfrom info such that it points to valid move source
within the working copy (and thus hopefully the same branch) instead.

* subversion/include/private/svn_wc_private.h
  (svn_wc__move_fixup): Declare.

* subversion/include/svn_client.h
  (svn_client_addremove): Fix name of first argument (cosmetic change).
  (svn_client_match_up_local_deletes_and_adds): Declare the new API.

* subversion/libsvn_client/addremove.c
  (addremove_status_baton): Add tables for added and deleted nodes.
  (addremove_status_func): Collect added and deleted nodles in addition to,
   or instead of, missing and unversioned ones.
  (suggest_moves): Rename parameters from unversioned/missing to added/deleted.
  (addremove): Stop moving items, just add or delete them.  
  (match_up_local_deletes_and_adds,
   svn_client_match_up_local_deletes_and_adds): New. Implement the new API.

* subversion/libsvn_wc/copy.c
  (svn_wc__move_fixup): Implement new private API at the libsvn_wc layer.

* subversion/libsvn_wc/wc-queries.sql
  (STMT_SET_MOVED_HERE_RECURSIVE): New query which sets 'moved_here'.

* subversion/libsvn_wc/wc_db.c
  (get_origin_of_delete_op_root): New helper function, factored out of ...
  (get_info_for_copy): ... this one.
  (set_moved_here_recursive, update_origin): New helper functions.
  (move_fixup, svn_wc__db_move_fixup): Implement wc_db support for new API.

* subversion/libsvn_wc/wc_db.h
  (svn_wc__db_move_fixup): Declare.

* subversion/svn/addremove-cmd.c
  (wrap_illegal_target_error): New helper, factored out of ...
  (svn_cl__addremove): ... this function. After adding unversioned and deleting
   missing nodes, call the new API to create local moves.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/addremove@1805813 13f79535-47bb-0310-9956-ffa450edef68